% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate}
\alias{calculate}
\title{Calculate}
\usage{
calculate(n = 500, incidence = 30, prob_skip_sick = 0, days_infectious = 10)
}
\arguments{
\item{n}{The number of participants}

\item{incidence}{Weekly incidence per 100,000}

\item{prob_skip_sick}{Likelihood of skipping of a person who is infected}

\item{days_infectious}{Number of days of infectiousness of a person}
}
\value{
A dataframe
}
\description{
Calculate the likelihood of attendance of infectious person
}
